update in 20220310

+++ Marlin/src/lcd/dwin/dwin_ui/dwin_comm.cpp
@@ -443,13 +443,13 @@ void ICON_YESorNO(uint8_t Option){
 	DWIN_FEEDBACK_TICK();
 }
 
-void DWIN_Show_Status_Message(const uint16_t color, char *string, const uint8_t show_seconds ){
+void DWIN_Show_Status_Message(const uint16_t color, char *string, const uint16_t show_seconds ){
 	Clear_Dwin_Area(AREA_BOTTOM);
 	dwinLCD.Draw_String(true, true, font8x16, color, COLOR_BG_BLACK, 10, 455, string);
 	set_status_bar_showtime(show_seconds);
 }
 
-void DWIN_Show_Status_Message(const uint16_t color, PGM_P string, const uint8_t show_seconds ){
+void DWIN_Show_Status_Message(const uint16_t color, PGM_P string, const uint16_t show_seconds ){
 	Clear_Dwin_Area(AREA_BOTTOM);
 	dwinLCD.Draw_String(true, true, font8x16, color, COLOR_BG_BLACK, 10, 455, string);
 	set_status_bar_showtime(show_seconds);
